Implemeted soft delete to all tables and added product delete mutation

+++ MundoBalloonApi.business/DataObjects/Entities/UserCart.cs
@@ -4,8 +4,8 @@ public class UserCart
 {
     public int UserId { get; set; }
     public string? Sku { get; set; }
-    public decimal Quantity { get; set; }
-    public decimal Price { get; set; }
+    public double Quantity { get; set; }
+    public double Price { get; set; }
     public int? ProductVariantId { get; set; }
     public ProductVariant? Variant { get; set; }
 }
