+++ MundoBalloonApi.infrastructure/Data/Models/OcassionCartDetail.cs
@@ -5,8 +5,8 @@ public class OcassionCartDetail : BaseEntity
     public int OccasionCartId { get; set; }
     public string? Sku { get; set; }
     public int ProductVariantId { get; set; }
-    public decimal Quantity { get; set; }
-    public decimal Price { get; set; }
+    public double Quantity { get; set; }
+    public double Price { get; set; }
     public string? Label { get; set; }
 
     public virtual OccasionCart? OccasionCart { get; set; }
