+++ MundoBalloonApi.infrastructure/Data/Models/UserCart.cs
@@ -4,8 +4,8 @@ public class UserCart : BaseEntity
 {
     public int UserId { get; set; }
     public string? Sku { get; set; }
-    public decimal Quantity { get; set; }
-    public decimal Price { get; set; }
+    public double Quantity { get; set; }
+    public double Price { get; set; }
     public int? ProductVariantId { get; set; }
 
     public virtual ProductVariant? ProductVariant { get; set; }
